[UPDATE]Se refactoriza el objeto ApiConstants y se crea el fragment de ultimas novedades para el home
