+++ app/src/main/java/com/melvin/ongandroid/view/fragments/LastestNewsHomeFragment.kt
@@ -1,0 +1,25 @@
+package com.melvin.ongandroid.view.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.melvin.ongandroid.R
+
+class LastestNewsHomeFragment : Fragment() {
+
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_lastest_news_home, container, false)
+    }
+}
